+++ packages/edit-site/src/components/post-edit/index.js
@@ -47,9 +47,9 @@ function PostEditForm( { postType, postId } ) {
 	const { fields } = usePostFields();
 	const form = {
 		type: 'panel',
-		fields: [ 'title', 'author' ],
+		fields: [ 'title', 'author', 'date' ],
 	};
-	const [ edits, setEdits ] = useState( {} );
+	const [ edits, setEdits ] = useState( initialEdits );
 	const itemWithEdits = useMemo( () => {
 		return {
 			...initialEdits,
@@ -71,7 +71,6 @@ function PostEditForm( { postType, postId } ) {
 				...edits,
 			} );
 		}
-		setEdits( {} );
 	};
 
 	const isUpdateDisabled = ! isItemValid( itemWithEdits, fields, form );